refactor TweetController with services and repositories layers
