+++ app/formats/TweetWithMemberView.scala
@@ -1,4 +1,4 @@
-package models
+package formats
 
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
